Removed the ability to delete a field anyway from the condition usage alert, moved field condition checking logic to condition.tsx
