+++ designer/src/components/field-editor.tsx
@@ -52,7 +52,7 @@ import {TakePhotoFieldEditor} from './Fields/TakePhotoField';
 import {TemplatedStringFieldEditor} from './Fields/TemplatedStringFieldEditor';
 import {TextFieldEditor} from './Fields/TextFieldEditor';
 import {useState, useMemo} from 'react';
-import {isFieldUsedInCondition} from './condition';
+import {findFieldCondtionUsage} from './condition';
 
 type FieldEditorProps = {
   fieldName: string;
@@ -101,7 +101,7 @@ export const FieldEditor = ({
   const deleteField = (evt: React.SyntheticEvent) => {
     evt.stopPropagation();
 
-    const usage = findFieldUsage(fieldName, allFields, allFviews);
+    const usage = findFieldCondtionUsage(fieldName, allFields, allFviews);
 
     if (usage.length > 0) {
       setConditionsAffected(usage);
@@ -112,15 +112,6 @@ export const FieldEditor = ({
         payload: {fieldName, viewId},
       });
     }
-  };
-
-  const confirmDelete = () => {
-    dispatch({
-      type: 'ui-specification/fieldDeleted',
-      payload: {fieldName, viewId},
-    });
-
-    setDeleteWarningOpen(false);
   };
 
   const getFieldLabel = () => {
@@ -174,48 +165,6 @@ export const FieldEditor = ({
       moveFieldCallback(targetViewId);
       handleCloseMoveDialog();
     }
-  };
-
-  /**
-   * Finds where a field is used in conditions or templated string fields
-   */
-  const findFieldUsage = (
-    fieldName: string,
-    allFields: Record<string, any>,
-    allFviews: Record<string, any>
-  ): string[] => {
-    const affected: string[] = [];
-
-    // Check section-level conditions
-    for (const sectionId in allFviews) {
-      const condition = allFviews[sectionId].condition;
-      if (isFieldUsedInCondition(condition, fieldName)) {
-        affected.push(`Section: ${allFviews[sectionId].label}`);
-      }
-    }
-
-    // Check field-level conditions
-    for (const fId in allFields) {
-      const condition = allFields[fId].condition;
-      if (isFieldUsedInCondition(condition, fieldName)) {
-        const label = allFields[fId]['component-parameters']?.label ?? fId;
-        affected.push(`Field Condition: ${label}`);
-      }
-    }
-
-    // Check for Templated String Fields using the deleted field
-    for (const fId in allFields) {
-      if (allFields[fId]['component-name'] === 'TemplatedStringField') {
-        const template = allFields[fId]['component-parameters']?.template || '';
-
-        if (template.includes(`{{${fieldName}}}`)) {
-          const label = allFields[fId]['component-parameters']?.label ?? fId;
-          affected.push(`Templated String: ${label} (uses '{{${fieldName}}}')`);
-        }
-      }
-    }
-
-    return affected;
   };
 
   // memoize the form value
@@ -401,18 +350,21 @@ export const FieldEditor = ({
           open={deleteWarningOpen}
           onClose={() => setDeleteWarningOpen(false)}
         >
-          <DialogTitle>Field Deletion Warning</DialogTitle>
+          <DialogTitle>Can Not Delete Field</DialogTitle>
           <DialogContent>
             <p>
-              This field is referenced in the following conditions. Deleting it
-              will invalidate them:
+              This field is referenced in the following{' '}
+              {conditionsAffected.length === 1 ? 'condition' : 'conditions'}:
             </p>
             <ul>
               {conditionsAffected.map((condition, index) => (
                 <li key={index}>{condition}</li>
               ))}
             </ul>
-            <p>Are you sure you want to proceed?</p>
+            <p>
+              Please delete {conditionsAffected.length === 1 ? 'it' : 'them'}{' '}
+              first to proceed.
+            </p>
           </DialogContent>
           <DialogActions>
             <Button
@@ -421,10 +373,7 @@ export const FieldEditor = ({
                 setDeleteWarningOpen(false);
               }}
             >
-              Cancel
-            </Button>
-            <Button onClick={confirmDelete} color="error">
-              Delete Anyway
+              Dismiss
             </Button>
           </DialogActions>
         </Dialog>
